+++ src/features/app/hooks/useAppState.ts
@@ -108,6 +108,7 @@ export const useAppState = (): AppState => {
   const [hideTrayIcon, setHideTrayIcon] = useState(false);
   const [hideDockIcon, setHideDockIcon] = useState(false);
   const [edgeDocking, setEdgeDocking] = useState(false);
+  const [checkUpdateOnStartup, setCheckUpdateOnStartup] = useState(true);
   const [customBackground, setCustomBackground] = useState<string>("");
   const [customBackgroundOpacity, setCustomBackgroundOpacity] = useState(45);
   const [surfaceOpacity, setSurfaceOpacity] = useState(50);
@@ -334,6 +335,8 @@ export const useAppState = (): AppState => {
     setHideDockIcon,
     edgeDocking,
     setEdgeDocking,
+    checkUpdateOnStartup,
+    setCheckUpdateOnStartup,
 
     customBackground,
     setCustomBackground,